+++ core/array_utils.h
@@ -54,60 +54,60 @@ static inline int string_bsearch_comp(const void* l, const void* r)
 }
 
 // Fills an array of integers with the given value.
-void int_fill(int* array, int length, int value);
+void int_fill(int* array, size_t length, int value);
 
 // Executes a linear search for an element in an unsorted array of integers, 
 // returning a pointer to the element if it's found and NULL if it's not.
-int* int_lsearch(int* array, int length, int element);
+int* int_lsearch(int* array, size_t length, int element);
 
 // Executes a binary search for an element in a sorted array of integers, 
 // returning a pointer to the element if it's found and NULL if it's not.
-int* int_bsearch(int* array, int length, int element);
+int* int_bsearch(int* array, size_t length, int element);
 
 // Returns the index of the (sorted) array at which the desired element 
 // appears (if it is present), or would appear (if it is not).
-int int_lower_bound(int* array, int length, int element);
+int int_lower_bound(int* array, size_t length, int element);
 
 // Sorts (in-place) the elements in an array of integers. Uses qsort().
-void int_qsort(int* array, int length);
+void int_qsort(int* array, size_t length);
 
 // Sorts (in-place) the elements in an array of integer pairs. Uses qsort().
-void int_pair_qsort(int* array, int length);
+void int_pair_qsort(int* array, size_t length);
 
 // Fills an array of indices with the given value.
-void index_fill(index_t* array, int length, index_t value);
+void index_fill(index_t* array, size_t length, index_t value);
 
 // Executes a linear search for an element in an unsorted array of indices, 
 // returning a pointer to the element if it's found and NULL if it's not.
-index_t* index_lsearch(index_t* array, int length, index_t element);
+index_t* index_lsearch(index_t* array, size_t length, index_t element);
 
 // Executes a binary search for an element in a sorted array of indices, 
 // returning a pointer to the element if it's found and NULL if it's not.
-index_t* index_bsearch(index_t* array, int length, index_t element);
+index_t* index_bsearch(index_t* array, size_t length, index_t element);
 
 // Returns the index of the (sorted) array at which the desired element 
 // appears (if it is present), or would appear (if it is not).
-int index_lower_bound(index_t* array, int length, index_t element);
+int index_lower_bound(index_t* array, size_t length, index_t element);
 
 // Sorts (in-place) the elements in an array of indices. Uses qsort().
-void index_qsort(index_t* array, int length);
+void index_qsort(index_t* array, size_t length);
 
 // Fills an array of real numbers with the given value.
-void real_fill(real_t* array, int length, real_t value);
+void real_fill(real_t* array, size_t length, real_t value);
 
 // Executes a linear search for an element in an unsorted array of real numbers, 
 // returning a pointer to the element if it's found and NULL if it's not.
-real_t* real_lsearch(real_t* array, int length, real_t element);
+real_t* real_lsearch(real_t* array, size_t length, real_t element);
 
 // Executes a binary search for an element in a sorted array of real numbers, 
 // returning a pointer to the element if it's found and NULL if it's not.
-real_t* real_bsearch(real_t* array, int length, real_t element);
+real_t* real_bsearch(real_t* array, size_t length, real_t element);
 
 // Returns the index of the (sorted) array at which the desired element 
 // appears (if it is present), or would appear (if it is not).
-int real_lower_bound(real_t* array, int length, real_t element);
+int real_lower_bound(real_t* array, size_t length, real_t element);
 
 // Sorts (in-place) the elements in an array of reals. Uses qsort().
-void real_qsort(real_t* array, int length);
+void real_qsort(real_t* array, size_t length);
 
 #endif
